Restrict untrusted local alloc limit to 1024.

+++ XstarS.GuidGenerators/Generators/GuidGeneratorExtensions.cs
@@ -77,7 +77,7 @@ public static class GuidGeneratorExtensions
         encoding ??= Encoding.UTF8;
 
         var nameLength = encoding.GetByteCount(name);
-        var nameBytes = (nameLength <= 1024) ?
+        var nameBytes = ((uint)nameLength <= 1024) ?
             (stackalloc byte[nameLength]) : (new byte[nameLength]);
         var bytesWritten = encoding.GetBytes(name, nameBytes);
         Debug.Assert(bytesWritten == nameLength);
@@ -147,7 +147,7 @@ public static class GuidGeneratorExtensions
         encoding ??= Encoding.UTF8;
 
         var nameLength = encoding.GetByteCount(name);
-        var nameBytes = (nameLength <= 1024) ?
+        var nameBytes = ((uint)nameLength <= 1024) ?
             (stackalloc byte[nameLength]) : (new byte[nameLength]);
         var bytesWritten = encoding.GetBytes(name, nameBytes);
         Debug.Assert(bytesWritten == nameLength);
